pyproject.toml
[tool.poetry]
name = "eora-test"
version = "0.1.0"
description = ""
authors = ["Viacheslav Beatum <podarkiprod@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytelegrambotapi = "^4.23.0"
python-dotenv = "^1.0.1"
aiohttp = "^3.10.6"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
